Allow closed tickets to be bookmarked
Previously, `ticketIsBookmarkable` got the ticket's status, defaulting
to CLOSED, and checked that against CLOSED. I have found a use-case for
bookmarking closed tickets, so I wanted to undo that check.
Unfortunately, that check had the (useful) byproduct of also making
not-yet-saved tickets not bookmarkable (in case someone adds it to the
New Ticket sidebar). Thus, I also check the ID to see whether it's null.